mvvm part 1 added
