+++ app/src/main/AndroidManifest.xml
@@ -12,10 +12,10 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".activities.Activator"/>
-        <activity android:name=".activities.MainActivity"/>
-        <activity android:name=".activities.MovieInfoActivity"/>
-        <activity android:name=".activities.SignInActivity">
+        <activity android:name=".view.activities.Activator"/>
+        <activity android:name=".view.activities.MainActivity"/>
+        <activity android:name=".view.activities.MovieInfoActivity"/>
+        <activity android:name=".view.activities.SignInActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
