+++ app/src/main/java/com/example/movie_db/model/data/movie/MovieResponse.kt
@@ -1,5 +1,6 @@
-package com.example.movie_db.classes
+package com.example.movie_db.model.data.movie
 
+import com.example.movie_db.model.data.movie.Movie
 import com.google.gson.annotations.SerializedName
 
 class MovieResponse {
