+++ app/src/main/java/com/example/movie_db/model/data/movie/SavingResponse.kt
@@ -1,4 +1,4 @@
-package com.example.movie_db.classes
+package com.example.movie_db.model.data.movie
 
 import com.google.gson.annotations.SerializedName
