+++ app/src/main/java/com/example/movie_db/model/database/MovieDatabase.kt
@@ -1,9 +1,10 @@
-package com.example.movie_db.classes
+package com.example.movie_db.model.database
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.example.movie_db.model.data.movie.Movie
 
 @Database(entities = [Movie::class], version = 1, exportSchema = false)
 abstract class MovieDatabase : RoomDatabase() {
