+++ app/src/main/java/com/example/movie_db/view/activities/SignInActivity.kt
@@ -1,4 +1,4 @@
-package com.example.movie_db.activities
+package com.example.movie_db.view.activities
 
 import android.content.Context
 import android.content.Intent
@@ -9,13 +9,13 @@ import android.util.Log
 import android.view.View
 import android.widget.*
 import com.example.movie_db.R
-import com.example.movie_db.Retrofit
+import com.example.movie_db.model.network.Retrofit
 import com.example.movie_db.BuildConfig
-import com.example.movie_db.classes.TokenResponse
-import com.example.movie_db.classes.UserResponse
-import com.example.movie_db.classes.User
-import com.example.movie_db.classes.LoginResponse
-import com.example.movie_db.classes.SessionResponse
+import com.example.movie_db.model.data.authentication.TokenResponse
+import com.example.movie_db.model.data.authentication.UserResponse
+import com.example.movie_db.model.data.authentication.User
+import com.example.movie_db.model.data.authentication.LoginResponse
+import com.example.movie_db.model.data.authentication.SessionResponse
 import com.google.gson.Gson
 import com.google.gson.JsonObject
 import android.widget.ProgressBar
