+++ app/src/main/java/com/example/movie_db/view/fragments/FragmentProfile.kt
@@ -1,4 +1,4 @@
-package com.example.movie_db.fragments
+package com.example.movie_db.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import com.example.movie_db.R
-import com.example.movie_db.classes.User
+import com.example.movie_db.model.data.authentication.User
 
 class FragmentProfile : Fragment() {
     override fun onCreateView(
