+++ app/src/main/res/layout/fragments_activity.xml
@@ -53,7 +53,7 @@
             android:background="@color/colorLight"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/toolbar"
-            tools:context=".activities.SignInActivity">
+            tools:context=".view.activities.SignInActivity">
 
             <include layout="@layout/movies_list" />
         </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
